requirements.txt
https://github.com/thiagopena/geraldo/archive/master.zip

Django==2.0.*
dj-database-url==0.5.*
python-decouple==3.1
pytz==2018.3
reportlab==3.4.*
six==1.11.*
cryptography==2.2.*
lxml==4.2.*
pyOpenSSL==17.5.*
mysqlclient==1.4.6
